fix(ringbuffer): use native size for find offsets
rbuf_find_bytes() scans up to the ringbuffer byte-count domain, but reported the found offset through an int pointer. On a ringbuffer whose readable data exceeds INT_MAX bytes, a full scan or late match could overflow the signed output counter even though the buffer itself is sized and indexed with size_t.

Current vs potential bug: current ACLK HTTP and WebSocket users keep buffers below INT_MAX, so this is not a known currently exercised production corruption path in those callers. The generic ringbuffer API nevertheless had a deterministic type-width mismatch for large buffers.

Solution: change rbuf_find_bytes() to report offsets through size_t *, update every direct in-repo caller to use size_t local indexes, and adjust WebSocket diagnostics to use size_t formatting. Focused unit coverage now verifies match and no-match offsets.

Functional impact: existing parsing behavior and return codes are unchanged. Offsets still represent bytes from the current tail; the output type now matches the rest of the ringbuffer capacity and byte-count API.

Performance impact: neutral. The scan loop still increments one pointer and one offset per byte; the offset type now uses the native size_t width already used by the surrounding ringbuffer code.

Validation: inspected the staged diff and all direct rbuf_find_bytes() call sites. Worker validation passed whitespace checks, changed object compilation, a local ringbuffer unittest runner, and reviewer quorum after rerunning an incomplete reviewer output; a broader netdata target stopped on an unrelated missing ACLK proto after changed objects compiled.

+++ src/libnetdata/ringbuffer/ringbuffer.c
@@ -226,7 +226,7 @@ int rbuf_memcmp_n(rbuf_t buffer, const char *to_cmp, size_t to_cmp_bytes)
     return rbuf_memcmp(buffer, buffer->tail, to_cmp, to_cmp_bytes);
 }
 
-char *rbuf_find_bytes(rbuf_t buffer, const char *needle, size_t needle_bytes, int *found_idx)
+char *rbuf_find_bytes(rbuf_t buffer, const char *needle, size_t needle_bytes, size_t *found_idx)
 {
     const char *ptr = buffer->tail;
     *found_idx = 0;
